Split dnsconfig.js into multiple files
