+++ globals/providers.js
@@ -1,0 +1,3 @@
+var REG_NONE = NewRegistrar('none');
+var REG_OVH = NewRegistrar('ovh');
+var DSP_DESEC = NewDnsProvider('desec');
